+++ modules/axfs/src/lib.rs
@@ -3,6 +3,7 @@
 pub mod block;
 pub mod devfs;
 pub mod fat32;
+pub mod ext4;
 pub mod memfs;
 pub mod mount;
 pub mod procfs;